replaced useState hook with useReducer hook in add recipe card
