+++ client/src/App.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 
 import MenuPanel from "./components/menu-panel/menu-panel.component";
-import RecipeCard from "./components/add-recipe/add-recipe.component";
+import RecipeCard from "./components/add-recipe/add-recipe-card.component";
 import { GlobalStyle } from "./global.styles";
 
 const MenuPage = () => <div>MENU ITEMS</div>;
