+++ client/src/components/add-recipe/add-recipe-card.styles.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+import ButtonRound from "../button-round/button-round.component";
+
+export const RecipeCardContainer = styled.div`
+  width: 900px;
+  height: 629px;
+  box-shadow: 0px 0px 25px rgba(0, 0, 0, 0.15);
+  margin: auto 10em auto auto;
+  display: flex;
+  flex-direction: column;
+  border-radius: 15px;
+`;
