+++ client/src/components/add-recipe/add-recipe-card.types.ts
@@ -1,0 +1,17 @@
+export interface IRecipeDetails {
+  icon: any;
+  recipeDetails: string;
+  units: string;
+  quantity: number;
+}
+
+export interface IState {
+  recipeDetails: IRecipeDetails[];
+  totalQuantity: number;
+  totalTime: number;
+}
+
+export interface IActions {
+  type: string;
+  payload: any;
+}
