+++ client/src/components/add-recipe/recipe-general-details/recipe-general.styles.tsx
@@ -1,15 +1,5 @@
 import styled from "styled-components";
-import ButtonRound from "../button-round/button-round.component";
-
-export const RecipeCardContainer = styled.div`
-  width: 900px;
-  height: 629px;
-  box-shadow: 0px 0px 25px rgba(0, 0, 0, 0.15);
-  margin: auto 10em auto auto;
-  display: flex;
-  flex-direction: column;
-  border-radius: 15px;
-`;
+import ButtonRound from "../../button-round/button-round.component";
 
 export const CardTop = styled.div`
   display: flex;
@@ -50,24 +40,4 @@ export const RecipeInputs = styled.ul`
     padding: 5px;
     border: 1px solid ${({ theme }) => theme.textContrast.default};
   }
-`;
-
-export const RecipeTable = styled.table`
-  width: 100%;
-  border-collapse: collapse;
-`;
-
-export const RecipeTableHeader = styled.tr`
-  background: ${({ theme }) => theme.background};
-  height: 40px;
-`;
-
-export const TableUpdateButtonsContainer = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-`;
-
-export const UpdateTableButton = styled(ButtonRound)`
-  border-radius: unset;
-  height: 40px;
 `;
